+++ src/eval/lang.rs
@@ -1,8 +1,5 @@
 use std::collections::HashMap;
 use std::fmt::Debug;
-use std::marker::PhantomData;
-
-use bellperson::Circuit;
 
 use crate::coprocessor::Coprocessor;
 use crate::field::LurkField;
@@ -10,9 +7,9 @@ use crate::store::{Ptr, Store};
 use crate::sym::Sym;
 
 // TODO: Define a trait for the Hash and parameterize on that also.
-#[derive(Debug, Clone)]
+#[derive(Debug, Default, Clone)]
 pub struct Lang<'a, F: LurkField> {
-    coprocessors: HashMap<Sym, Box<&'a dyn Coprocessor<F>>>,
+    coprocessors: HashMap<Sym, &'a dyn Coprocessor<F>>,
 }
 
 impl<'a, F: LurkField> Lang<'a, F> {
@@ -23,16 +20,15 @@ impl<'a, F: LurkField> Lang<'a, F> {
     }
 
     pub fn add_coprocessor(&mut self, name: Sym, cproc: &'a dyn Coprocessor<F>) {
-        self.coprocessors.insert(name, Box::new(cproc));
+        self.coprocessors.insert(name, cproc);
     }
 
     pub fn lookup(&self, s: &Store<F>, name: Ptr<F>) -> Option<&dyn Coprocessor<F>> {
         let name_ptr = s.fetch_maybe_sym(&name);
 
         name_ptr
-            .as_ref()
-            .and_then(|sym| self.coprocessors.get(sym))
-            .map(|x| **x)
+            .and_then(|sym| self.coprocessors.get(&sym))
+            .copied()
     }
 }
