+++ src/eval/reduction.rs
@@ -54,6 +54,7 @@ impl<F: LurkField> Control<F> {
     }
 }
 
+#[allow(clippy::too_many_arguments)]
 fn reduce_with_witness_inner<F: LurkField>(
     expr: Ptr<F>,
     env: Ptr<F>,
